requirements.txt
pyserial
matplotlib
pillow
pystray
pyinstaller
pythonnet